DDF-2861 Rename routemanager
