+++ broker/broker-route-manager/pom.xml
@@ -25,7 +25,7 @@
     <name>DDF :: Broker :: Route Manager</name>
     <packaging>bundle</packaging>
 
-    <artifactId>broker-routemanager</artifactId>
+    <artifactId>broker-route-manager</artifactId>
 
     <dependencies>
         <dependency>
