+++ broker/broker-route-manager/src/test/java/org/codice/ddf/broker/route/manager/DynamicRouteDeployerTest.java
@@ -11,7 +11,7 @@
  * is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.broker.routemanager;
+package org.codice.ddf.broker.route.manager;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
