+++ broker/pom.xml
@@ -121,7 +121,7 @@
         <module>broker-connection-factory</module>
         <module>camel-sjms2</module>
         <module>broker-undelivered-messages-ui</module>
-        <module>broker-routemanager</module>
+        <module>broker-route-manager</module>
         <module>broker-app</module>
     </modules>
 </project>
